Revert "Js2 ca engelund joran"
