--- post-filter.html
@@ -1,102 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta
-      name="description"
-      content="home page, share and read other posts in the feed, search on people, posts and contacts on SoMe "
-    />
-    <link rel="icon" type="image/x-icon" href="/assets/images/favicon.svg" />
-    <link rel="stylesheet" href="/dist/css/styles.css" />
-    <title>Home | SoMe</title>
-  </head>
-  <body class="bg-secondary d-flex flex-column min-vh-100">
-    <header class="sticky-top">
-      <nav class="navbar navbar-expand-md shadow bg-light">
-        <div class="container-fluid">
-          <a href="/home.html" class="navbar-brand"
-            ><img
-              src="/assets/images/SoMe.svg"
-              alt="logo of SoMe"
-              class="img-logo"
-          /></a>
-          <button
-            type="button"
-            class="navbar-toggler"
-            data-bs-toggle="collapse"
-            data-bs-target="#navbarSupportedContent"
-            aria-controls="navbarSupportedContent"
-            aria-expanded="false"
-            aria-label="Toggle navigation"
-          >
-            <span class="navbar-toggler-icon"></span>
-          </button>
-          <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <ul class="navbar-nav me-auto">
-              <li class="nav-item">
-                <a
-                  href="/home.html"
-                  class="nav-link bg-primary text-light rounded py-1 px-3 mx-2 shadow"
-                  >Home</a
-                >
-              </li>
-              <li class="nav-item">
-                <a href="/profile.html" class="nav-link rounded py-1 px-3 mx-2"
-                  >Profile</a
-                >
-              </li>
-              <li class="nav-item">
-                <a
-                  href="/index.html"
-                  class="nav-link rounded py-1 px-3 mx-2 sign-out"
-                  >Sign out</a
-                >
-              </li>
-            </ul>
-            <form class="d-flex" role="search">
-              <input
-                type="search"
-                class="form-control search-bar ps-4 navbar-dark"
-                placeholder="Search"
-                aria-label="search"
-              />
-            </form>
-          </div>
-        </div>
-      </nav>
-    </header>
-    <main>
-      <section id="feed-container">
-        <div class="container">
-          <div class="row justify-content-center">
-            <div
-              id="post-container"
-              class="col-12 lg-6 d-flex flex-column justify-content-center align-items-center text-left"
-            >
-              <h1 class="custom-label mt-3 tags-header"></h1>
-              <div class="d-flex flex-row gap-3">
-                <a
-                  href="/home.html"
-                  class="btn btn-primary mt-4 mb-3 px-5 text-center btn-custom shadow"
-                  >Back to feed</a
-                >
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-    </main>
-    <footer class="bg-light mt-auto p-2 text-center">
-      <a
-        target="_blank"
-        href="https://joranengelund.github.io/"
-        class="nav-item"
-        >All rights reserved © JayDesign 2022</a
-      >
-    </footer>
-    <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="/src/js/index.mjs" type="module"></script>
-  </body>
-</html>
