+++ profile.html
@@ -48,9 +48,7 @@
                 >
               </li>
               <li class="nav-item">
-                <a
-                  href="/index.html"
-                  class="nav-link rounded py-1 px-3 mx-2 sign-out"
+                <a href="/index.html" class="nav-link rounded py-1 px-3 mx-2"
                   >Sign out</a
                 >
               </li>
