+++ src/js/posts/render.mjs
@@ -1,5 +1,4 @@
 import { createPosts } from "./templates/createPost.mjs";
-import { filterTemplateTags } from "../filter/templates/filter-dropdown.mjs";
 
 /**
  * // Loops over all posts with forEach method and destructures it to access necessary properties
@@ -28,7 +27,6 @@ export function renderPost(posts) {
     }) => {
       const { name, avatar } = author;
       const { comments: commentCount, reactions: reactionCount } = _count;
-      filterTemplateTags(tags);
       createPosts(
         title,
         created,
