+++ src/scss/styles.scss
@@ -106,7 +106,7 @@ footer a,
 }
 
 .card-img-bottom {
-  max-width: 100%;
+  max-width: 40%;
   height: auto;
 }
 
@@ -136,11 +136,6 @@ footer a,
     border-radius: 50%;
     box-shadow: $box-shadow;
   }
-}
-
-.tags {
-  max-height: 11rem;
-  overflow: hidden scroll;
 }
 
 @media only screen and (max-width: 400px) {
